functional tests: split code according to features
The Test.cpp file was getting big; let's split it and group tests by features
or requirements.
