+++ test/functional-tests/CMakeLists.txt
@@ -41,7 +41,10 @@ if(BUILD_TESTING)
     include_directories("${CMAKE_CURRENT_SOURCE_DIR}/include")
 
     # Add unit test
-    add_executable(parameterFunctionalTest Test.cpp catch.cpp)
+    add_executable(parameterFunctionalTest
+                   Basic.cpp
+                   FloatingPoint.cpp
+                   catch.cpp)
     set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++11")
 
     target_link_libraries(parameterFunctionalTest parameter)
